Add check for !https to mixed content IP address console message
MixedContentChecker::ShouldAutoupgrade currently adds a console message
unconditionally when there is a subresource on an https site if the
subresource URL is an IP address, regardless of whether it is mixed
content (i.e. even if the subresource is also on https). This adds a
check so the warning is not added if the subresource is also https.

Bug: 1454787
Change-Id: I34f0ab97c2046572ab89d380ab7d4219d283ab89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4692025
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1173127}
NOKEYCHECK=True
GitOrigin-RevId: b84698363d958ff7c0a2c5af342a65e2647dd700