[wasm] Simplify handling of streaming errors
This reverts an attempt to fix DCHECK errors we saw in the wild. The
actual fix was https://crrev.com/c/4654310.

As documented on {BeginRead} and {EndRead}, the {kErrored} status should
only be returned when in an errored state. In closed state, {kDone} will
be returned.

R=ahaas@chromium.org

Bug: 1449546
Change-Id: I23dcdeeb573a0f59627220f0c7a5c19c87c4e79d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4694102
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1173396}
NOKEYCHECK=True
GitOrigin-RevId: 372cea4b6d27d24ee30976722c6caa8f2abe5252