Add attribution-src-token to request
This CL is part of work needed to support background attributions
registrations by adding a hook in the KeepAliveUrlLoader.

See go/ara-keepalive-dd for details.

Alongside a navigation which wants (and is eligible) to register
attribution, we might independently send background attribution
registration requests. To process their responses, we need metadata
on the navigation which is only available in the browser.

Today, to get this metadata, we tag the navigation with an
`attribution-src-token`. When sending a background request, we
use `RegisterNavigationDataHost` to link the background request to the
token. We then observe navigations and tie the background request to the
navigation. https://source.chromium.org/chromium/chromium/src/+/main:content/browser/attribution_reporting/attribution_data_host_manager_impl.cc;l=593;bpv=0;bpt=0

With the new design, we have a hook in the KeepAliveUrlLoader instead
of using a pipe from blink. We can therefore no longer rely on the pipe
to pass down the `attribution_src_token`. When processing the request
we need to know the navigation to which it is "linked".

This is why, in this CL, we add a new optional `attribution-src-token`
on the request. See example usage in proof-of-concept CL here:
https://chromium-review.googlesource.com/c/chromium/src/+/4591551/16/content/browser/loader/keep_alive_url_loader_service.cc#226

Bug: 1374121
Change-Id: I6f48b1bfc7e23de4b777e6e9d64ff0333750d14a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4706804
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1174328}
NOKEYCHECK=True
GitOrigin-RevId: 767e677422c6af7597079af67f284cc0ef5df4b3