Remove min overscroll event delta check
There doesn't seem to be spec describing how the deltas should be set
and Safari and Firefox don't seem to support the overscroll event yet
(per my manual test using this codepen[1] and wpt.fyi[2]) so there
doesn't really seem to be a reason to enforce a check on the minimum
delta in each and every overscroll event.
For example, before this patch, if there was extra scrolling input
of 300px causing overscroll events with with individiual deltas of 48, 102, and 150, we could fire overscroll event with cumulative deltas of 48, 150 and 300 the test would fail because the first event had 48 (<50). The events report the cumulative deltas[3], not the individual values.

It does seem reasonable to keep the check that the maximum overscroll
is reasonably close to and not greater than the excess amount of
scroll input.

[1]https://codepen.io/awogbemila/pen/eYQVWrL
[2]https://wpt.fyi/results/dom/events/scrolling/overscroll-deltas.html?label=experimental&label=master&aligned
[3]https://source.chromium.org/chromium/chromium/src/+/71d03067cf95b8c8f42fd74c7334654e2a39639f:cc/input/input_handler.cc;l=1865

Bug: 1430215
Change-Id: I891711f346d6f0cd8c3b423c0a2e8d141c6dec6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4690405
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1174490}
NOKEYCHECK=True
GitOrigin-RevId: 74722528bfa501bb512ab848b3f158d7225ccfdf