[Private Network Access] Use initiator policies for nested navigations.
This change fixes a longstanding bug in the PNA implementation for
iframes. PNA checks apply to the fetch client's policies, not the
parent frame's policies.

Also introduce a couple new feature flags to control the rollout of
PNA for iframes: warning-only mode and full enforcement. These replace
the previous feature flag, which only applied to non-secure contexts.

The feature flag checking logic is moved to live inside
`DerivePrivateNetworkRequestPolicy()`, as is the case for worker flags.

A couple virtual test suites are introduced to exercise the code paths
behind the two new flags. These show that there remain bugs when the
feature is entirely enabled: preflight requests do not succeed. This
seems to be due to a strange interaction of CORS logic with PNA, where
navigation requests are made in no-cors mode with credentials. Further
work is needed to fix this, either by fixing the tests or by fixing
CORS logic.

Bug: chromium:1170335
Change-Id: Iea67b39c46ff48edac3f51e9ccc63d4bac47b7e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3945088
Auto-Submit: Titouan Rigoudy <titouan@chromium.org>
Reviewed-by: Jonathan Hao <phao@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1174883}
NOKEYCHECK=True
GitOrigin-RevId: ff5a84b3bba592f27fc27e4c447e9e5c3fa9298a