[layout] Don't mess with margins for alignment/RTL purposes.
Previously in block layout we'd adjust margins for text-align purposes
and for obtaining the correct line-left offset for RTL.

This resulted in the margins we'd store to be incorrect. Instead of
adjusting the margins account for alignment/RTL with an additional
offset instead.

Change-Id: Iaffbfcbc8780c84a86ef6ed0125d289d886a6145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4712468
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1175073}
NOKEYCHECK=True
GitOrigin-RevId: 8b22453afaafdc1b390bf31c9bd29e3a45e70e5d