Make ExceptionState usable during the type conversion.
It is suggested to use URLPattern::Create to parse URLPattern in
InstallEvent::registerRouter.  However, since we used mojo::ConvertTo,
which convert a type to the other type, I did not think it good fit to
pass ExceptionState there as-is.

This changed the interface name a bit and allow us to send the
ExceptionState so that we can use URLPattern::Create there.
Also, we can raise exception there.

Bug: 1457267
Change-Id: Id99879d5af47c6d5b689a38067e6e24f6cc10305
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4714470
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1175273}
NOKEYCHECK=True
GitOrigin-RevId: 3c1ac098515232e77390385c0d181640cc0274fa