[WebGL] Remove WEBGL_webcodecs_video_frame extension
This extension was an early prototype what became WebGPU's ability to
import VideoFrames in zero-copy fashion. It proved too difficult to use,
and its spec is currently marked as "Do Not Implement" [1]. This CL
removes its implementation from Chromium and instead makes the extension
unsupported in Chromium. The concrete motivation for doing this now (in
addition to code cleanup) is that its per-texture accesses for each
separate plane of a multiplanar SharedImage are problematic for our
introduction of Multiplanar SharedImage (which has only one texture for
the entire image).

Bug: 1462924
Change-Id: I08ace73456843c1b53ffc378bbf6e6c3adb446e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4698256
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1175440}
NOKEYCHECK=True
GitOrigin-RevId: 588ad3b5a9ca4a94d843fee2e68a66bfa995129a