Fix rSAFor WPT dual-write test case
The purpose of this test is to verify that a successful
requestStorageAccessFor() call also sets a storage-access grant so
that requestStorageAccess() should be resolved without a user gesture.
However the test was testing this behavior in a same-origin iframe,
in which case requestStorageAccess() is resolved not due to the
existing grant.

Change-Id: Ic16702271d1bed1f4faaacdd3ebef0dc152fecb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4721492
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Commit-Queue: Chris Fredrickson <cfredric@chromium.org>
Auto-Submit: Shuran Huang <shuuran@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1175699}
NOKEYCHECK=True
GitOrigin-RevId: 2dbe71691e30d88ea651b06e55e8f2b899e39aad