Fix <a>'s href getter to return the attribute value for invalid URL
The CL is a part of the URL Interop 2023.

The anchor element <a>'s href should return the content attribute
value as is if the invalid URL is set in its setter.

For example, Chromium fails the following test because `a.href`
returns "http://%00/#a", instead of "http://%00#a".

    const a = document.createElement("a");
    a.href = "http://%00#a";
    assert_equals(a.href, "http://%00#a");

This behavior is non-compliant with the standard [1].

There is a WIP CL [2] which attempts to fix this issue. The difference
between the previous one and this CL is:

- The flag is now Blink's RuntimeEnabledFeature
- The flag is enabled by default. This CL doesn't add VirtualTests.

Since the flag is now enabled by default, we have to rewrite
Chromium's internal web_tests which rely on the current non-compliant
behavior.

Since `a.href` is used internally as an implementation of
canonicalize() function [3] in web_tests, we have to rewrite many
URL's canonicalization web_tests defined mostly in fast/url/.

In contrast, we don't need to rewrite any WPT tests. They are all
correct.

As discussed in the previous CL, we don't send I2S for this change.

Rebaselines:

This CL includes many rebaselines. [4] is a diff before adding
rebaselines, which might be helpful to understand the main change.

Most rebaselines are understandable. Here are some notes for the record.

a-element-origin-expected.txt:

> FAIL Parsing origin: <x> against <sc://ñ> assert_equals: origin
> expected "null" but got ""

This is due to https://crbug.com/1416022.

The reason of PASS previously is that Chromium has two bugs which
*accidentally* made this test PASS.

Previously:

1. Chrome fails to parse <x> against <sc://ñ> (1st issue)
2. anchor.Url() wrongly returns invalid URL as is (2nd issue)
3. anchor.origin return null for the URL.

After this CL:

1. Chrome fails to parse <x> against <sc://ñ> (1st issue remains)
2. anchor.Url() returns null URL (after this CL)
3. anchor.origin return empty "" for null URL

Expected behavior [5]:

1. Chrome parse <x> against <sc://ñ> successfully
2. anchor.Url() returns its URL
3. anchor.origin return null for the URL

- [1] https://html.spec.whatwg.org/multipage/links.html#dom-hyperlink-href
- [2] https://crrev.com/c/4379704
- [3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/fast/url/resources/utilities.js
- [4] https://chromium-review.googlesource.com/c/chromium/src/+/4659525/0..8
- [5] https://html.spec.whatwg.org/multipage/links.html#dom-hyperlink-origin

Bug: 1415986
Change-Id: Id0397a66f614f32757c7d19771cd694a9ba03556
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4659525
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176367}
NOKEYCHECK=True
GitOrigin-RevId: a9232586feedbc5c4b97dd5e2c8a115edec35921