Make mojo and DB supports other fields in URLPattern.
This CL adds support on all other URLPattern fields to make it usable in
matching in the browser side process.  Also, this CL adds the
URLPatternOptions field so that some options set in `URLPattern`
constructor can be passed.  Especially, the ignore_case option has been
supported.

Also, we noticed that the empty field in the URLPattern means
the pattern should not match the URL with the field.
e.g. URLPattern({search:""}) should not match the URL with a search
string.
Database automatically absorb the difference by setting full wildcard
pattern to the fields other than pathname if the legacy pathname is
used.

Note that ServiceWorkerRouterEvaluator implementation and IDL
implementation will come in the other CLs later.

Minor change:
- renamed function and variable names in SW DB because we now handle not
  only pathnames but also other fields.

Bug: 1371756, 1467845
Change-Id: Ia94d8db9efe1ec37f2c34400a34dc239467d7ad2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4713808
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176509}
NOKEYCHECK=True
GitOrigin-RevId: 18196c6a774d2d25f45ce31a8e3bb7225e72943a