Clear media licenses along with dom storage types.
This ensures that [Clear-Site-Data: "storage"] will delete all buckets
and their contents. This is necessary because media licenses are
currently counted as a storage type; if they are not included in
deletion then the buckets themselves will not be deleted.

This is a temporary hack until crbug.com/1434517 is addressed.

Bug: 1422663,1434517
Change-Id: Iac88174a874a92dbaeea874ee67c783f9d3d7649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4702765
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176729}
NOKEYCHECK=True
GitOrigin-RevId: 2d7667b7299046aa573a4b551c0fb6a690ee307a