IndexedDB: save a memory allocation when not compressing values
When an IDB value doesn't compress much, we leave it uncompressed and
discard the memory that was allocated to hold the compressed version.
Then later on we allocate new memory to store a copy of the
uncompressed value.

This change keeps the memory that was allocated for the compressed
bytes and uses it for the uncompressed bytes, thus decrementing by
one the number of allocations/deallocations.

Bug: 1459300
Change-Id: I711666ba1de60ef95e18351a04072796c5416862
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4726724
Reviewed-by: Austin Sullivan <asully@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1176762}
NOKEYCHECK=True
GitOrigin-RevId: 91a623a5bebf7cb65b81cf3164fea217151da554