[EditContext] End composition when empty SetComposition received
When a composition is active and the user hits Escape, the
composition should be ended, the text in the composition range should
be deleted, and compositionend should be fired. The signal to do this for InputMethodController is receiving an empty string in
SetComposition.

EditContext deletes the text in the composition range but does not
properly terminate the composition in this case.

In this CL, terminate the composition and fire compositionend when
receiving an empty string in SetComposition, to match the
behavior or InputMethodController and properly close the composition
when the user hits Escape.

Bug: 999184, 1468436
Change-Id: Ia496988a3da23ebf98e39d9fea33d382291bb213
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4729288
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1177364}
NOKEYCHECK=True
GitOrigin-RevId: ade4e237d7c1f6dee986399f417ee8ab226ccca5