Fenced frames: add WPT for scrollIntoView() bubbling.
This CL adds an external WPT to ensure that calling scrollIntoView()
inside a fenced frame does not cause its embedder to scroll as well.

This is needed for our spec work. We currently have the
`ScrollIntoViewFencedFrameBrowserTest` browser test, but we need to
create a version that can be added to our spec.

Change-Id: Ic970b7f517ef76084e984afcca475e808c86b51e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4739596
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Commit-Queue: Liam Brady <lbrady@google.com>
Cr-Commit-Position: refs/heads/main@{#1178080}
NOKEYCHECK=True
GitOrigin-RevId: 3c8eb328fc354a1ce0ad51c6a87e66b21fd9adf9