[LCPP] Always set LCPP hint
When we are committing the same RenderFrame as the previous document and
there's a pre-existing LCPP hint, and when there are no LCPP element
hints, we want to remove the existing hint. Hence this CL updates
SetLCPPHint() to clear the existing hint.

Bug: 1419756
Change-Id: Iae4735a19816ff47f74a05ef114e1c9157014ffd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4741252
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178364}
NOKEYCHECK=True
GitOrigin-RevId: 45fff1a9f70ca5296a04c2eb2cdd382cee8f07d1