Add WPT to test rounding for adCost
Create a new web platform test to test if adCost is rounded correctly
after returned by generateBid(). The idea is to check the adCost from
browserSignals in ReportWin() and use sendReportTo() to verify if the
rounded value are in expected URL arrays.

Bug: 1425952
Change-Id: Id042f9df9a8e78a937e0cea3396f537888fc840b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4727328
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Tianyang Xu <xtlsheep@google.com>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178405}
NOKEYCHECK=True
GitOrigin-RevId: 67e4f2113685c386755f87a4248b0745f77968f8