[CodeHealth] Unify attribute maps in `PointerEventFactory`.
To be able to find the pointer-id for a `click` event for which the
corresponding pointer could be gone from PointerEventFactory, we need
to (temporarily) store removed pointers in a separate list.  Before
adding the separate list (in a follow-up CL), this CL makes a no-op
change to unify all pointer attributes (including positions) into a
single map keyed by pointer-id.
- Rename `pointer_id_mapping_` to `pointer_id_to_attributes_`.
- Move pointer position data in maps `*last_position_mapping_` to
  within `pointer_id_to_attributes_` map.
- Rewrite `typedef struct X {} X` to C++ style `struct X {}`.
- Add missing braces in multi-line `if`s.

Bug: 1264930
Change-Id: I3d98b426bafab175bce3547116372c744689cb48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4702190
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178501}
NOKEYCHECK=True
GitOrigin-RevId: d900a064457b02358b3962bb82c5ec3c4a2801c9