IndexedDB: remove an extra trip to task queue when firing result events.
See bug for overview of problem and performance results of this fix.

This patch removes the extra async step when returning a result via
IDBRequest. Results are returned via event dispatch, so this is
accomplished by removing the use of an EventQueue.

The IndexedDB spec often requires algorithms to “queue a task” which
forces asynchronocity. It also requires that “Requests must be executed
in the order in which they were made against the transaction.” IDB
results are returned via event dispatch. To enforce these two
requirements of asynchronicty and ordering, the events were always added
to an EventQueue. This is generally not necessary because:

* User actions that round trip through the browser process are already
  asynchronous (and responses are handled on the appropriate task
  queue)
* Browser-side logic processes and returns requests in the order in
  which they are created. Later, in Blink, IDBTransaction keeps an
  ordered list of returned results (a list of IDBRequestQueueItems)
  as they are processed. Thus ordering is maintained.

The meat of this change is the removal of various `QueueEvent` methods
and replacement with direct event dispatch. In the cases where there is
not already a round trip through the browser process (such as
`cursor.advance()` when values have been prefetched), or an algorithm
explicitly requires some steps be synchronous while other steps are
async (such as `transaction.abort()`, which immediately cancels all
requests but asynchronously resets requests' attributes and dispatches
events), we explicitly post tasks to the appropriate queue.

The IDB spec interleaves synchronous and asynchronous steps in its
algorithms. For example, `cursor.advance()` will immediately set the
request’s `readyState` to ‘pending’ but asynchronously dispatch the
result. There are many WPT that verify event ordering and other details
of these algorithms which indicate that these changes are safe.

Along with these functional changes comes some slight refactoring and
renaming.
* The term “Enqueue” was heavily overloaded, and is often no longer
  accurate, so it is changed to "Send" in a bunch of places.
* For clarity, function names generally become more unique.
* Some virtuals that were no longer necessary were removed
  (IDBFactoryClient knows its request is an IDBOpenRequest)
* For simplicity and consistency, requests are always sent to their
  transaction’s result_list_, even if they don’t require async
  processing.

Fixed: 965136

Change-Id: I107459148ce8253b16507095f8d22ccc737c4df1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4707660
Reviewed-by: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178509}
NOKEYCHECK=True
GitOrigin-RevId: 8ef880b9c7dc8f4b862d3354dbbfb852f7b35b0e