Clean up handling of incomplete relations
An incomplete relation is one where the target id is not yet on a
connected node in the DOM. In these cases, the source node often must
be serialized without the relation information. It's important to
reserialize the source of the relation with the corrected relation data once the relation target is known.

Bug: none
Change-Id: Ideb176ee55560ae46f9248ecaf47191a46b67264
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4739902
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178561}
NOKEYCHECK=True
GitOrigin-RevId: 7e6ac8a55866b4329d87e77d4ac5502d5afeba91