Support (behind a flag) any CSS display on <details> and <summary>.
This adds a new DetailsStyling flag (and virtual test suite) that is
intended as a feature flag for a set of changes for improving the
ability to style the <details> and <summary> elements.

It also adds the first change to that flag, which allows the <details>
and <summary> elements to take any value of the CSS display property
(instead of forcing <details> to be a block or list item, and <summary>
to be one of a slightly larger but still restricted list).

Bug: 1469418
Change-Id: Id9482b898d11d8059ac8d9ea2d63bb8a995cae6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4742959
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: David Baron <dbaron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178779}
NOKEYCHECK=True
GitOrigin-RevId: 6223abaa9f049345884ec0798573f16dab9400b3