Treat Settings::GetPrintingMaximumShrinkFactor() as-is.
We used to divide it by 1.33333333 by "accident", because the printing
code used the 1.33333333 factor as an encryption key, to make everything
hard to understand, prior to CL:4567866.

Adjust the default value of the printingMaximumShrinkFactor setting
accordingly, so that there won't be any behavior changes unless that
value is overridden. If the value *is* overridden, on the other hand
(this should be rare), the user may now be surprised to see that the
value specified will actually be respected, rather than be divided by
1.3333333. But hopefully this is acceptable.

Bug: 1296099
Change-Id: I376aeafb269f69e8bc15fd4751dfa171f6926037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4741715
Reviewed-by: Alice Boxhall <alice@igalia.com>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178881}
NOKEYCHECK=True
GitOrigin-RevId: 769fe978bebda6adfc0e50433f47c57220eda865