[gc] Apply MPC and inheritance directly in ComputedStyle's constructor.
Currently, when creating a new ComputedStyle object, we would be
first setting provisional subgroup pointers (based on the UA initial
style), and then overwrite parts of it using inheritance and MPC data.
For instance, for the worst-case situation of a partial MPC hit,
we could see something like this:

 1. Construct the object; increase the refcount of all subgroups.
 2. Call InheritFrom() with the parent; decrease the refcount of
    all the subgroups we used to have, increase the refcount of
    all the newly copied subgroups.
 3. Look up in the MPC, and find the partial hit.
 4. Call InheritFrom() with the returned style from MPC; decrease
    the refcount of all the subgroups _again_, and instead increase
    the refcount of the newly copied ones.

While the above is written in terms of refcounts, the motivation for
this patch is to make it cheaper to move towards having ComputedStyle
on Oilpan, where we have a very similar problem with subobject Member<>
pointer changes creating GC barriers (which can hit a slow path) --
with the important exception that no barriers can happen if the pointer
is set _during the constructor_ (because the ComputedStyle does not
exist yet, so it could not possibly be in the process of being traced
by a background thread). Thus, we want to move as much as possible of
this work into the constructor, setting the subgroup pointers once
there and then be done with it. (Of course, any later Set* would be
modifying the pointers as part of copy-on-write; this is something we
can only fix by moving fully towards a build-once pattern, which this
patch does not attempt.)

Thus, we replace InheritFrom() and CopyNonInheritedFromCached()
with a completely new constructor patterns, which takes in styles
to take inherited and non-inherited subgroups from, and merge the
subgroups in the constructors as needed. (The only real edge case
here is the SVG subgroup, which contains a mix of inherited and
non-inherited properties, so we need to merge that recursively
in the same way.) This means we cannot construct a ComputedStyle
until after the MPC lookup has happened -- this also means that
a lot of the old code which assumed a provisional ComputedStyle
existed (sometimes with certain properties already being set,
sometimes just assuming they were at the UA initial values)
had to be cleaned up, which has happened in some of the previous
patches.

This also fixes the issue where reset_on_new_style (formerly
custom_copy) fields in raredata groups would not be correctly
reset when creating a new style.

Style perftest (Zen 2, LTO but no PGO):

Initial style (µs)     Before     After    Perf      95% CI (BCa)
=================== ========= ========= ======= =================
ECommerce               11160     10995   +1.5%  [ +1.3%,  +1.8%]
Encyclopedia            77967     77315   +0.8%  [ +0.6%,  +1.1%]
Extension              126145    123744   +1.9%  [ +1.7%,  +2.2%]
News                    37288     36640   +1.8%  [ +1.6%,  +2.0%]
Search                  10988     10864   +1.1%  [ +0.9%,  +1.5%]
Social1                 22368     22084   +1.3%  [ +1.0%,  +1.5%]
Social2                 15647     15331   +2.1%  [ +1.9%,  +2.2%]
Sports                  44202     43669   +1.2%  [ +1.0%,  +1.5%]
Video                   31876     31354   +1.7%  [ +1.5%,  +1.8%]
Geometric mean                            +1.5%  [ +1.3%,  +1.7%]

Recalc style (µs)      Before     After    Perf      95% CI (BCa)
=================== ========= ========= ======= =================
ECommerce               10159      9924   +2.4%  [ +1.9%,  +2.8%]
Encyclopedia            58436     57444   +1.7%  [ +1.3%,  +2.2%]
Extension              114612    112627   +1.8%  [ +1.3%,  +2.2%]
News                    26704     26027   +2.6%  [ +2.2%,  +3.0%]
Search                   5653      5527   +2.3%  [ +2.0%,  +2.8%]
Social1                 15326     15024   +2.0%  [ +1.6%,  +2.3%]
Social2                 12063     11802   +2.2%  [ +1.9%,  +2.5%]
Sports                  25569     26106   -2.1%  [ -2.3%,  -1.7%]
Video                   20498     19927   +2.9%  [ +2.5%,  +3.1%]
Geometric mean                            +1.7%  [ +1.5%,  +2.1%]

MotionMark appears to be neutral.

Change-Id: Iede777c627901cf78dce097093c14a25be9cd718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4667195
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178967}
NOKEYCHECK=True
GitOrigin-RevId: 1607affb041293745538371d574db82032b947b3