[SDA] Fix crash when resolving inset with an unattached subject
When resolving a length from an inset, we use a length resolver, which
in turn requires a reference to the subject's style. When the subject
is not attached to the document, its style is not calculated and we
had a null deref. To avoid this problem, we verify that the subject
has a style before using the length converter. We handle the trivial
cases of a literal px or percent inset, and otherwise fall back to
marking the inset as being style dependent, in which case, it needs to
be resolved on each style update.

Bug: 1428965
Change-Id: I870d20a72d6dfd9cb787132022545442150d1ecb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4738874
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1178978}
NOKEYCHECK=True
GitOrigin-RevId: c688b6df90485a2330cb92b2686fc4647236dff8