Only add trailing border/padding to the right fragment, part 2.
I tried to do this in CL:4345873, but that didn't cover all the cases.
If the node being laid out is a formatting context root, floats may also
play a role, so just looking for in-flow breaks inside isn't sufficient
in such cases.

When printing, adding trailing border/padding incorrectly caused
monolithic overflow. If there was something tall monolithic inside as
well, we'd keep on breaking before it indefinitely, because the
monolithic overflow on the container provided a break opportunity.

Bug: 1468664
Change-Id: I0e0c8c47deba4ffa611157497bc639f3c268836f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4746460
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1179134}
NOKEYCHECK=True
GitOrigin-RevId: fcbc126a830b696ad7f7080f8474e8e3145e6676