[FedCM] Don't omit schemes when formatting URLs
We use the result of this function to construct a GURL, so
we can't omit any schemes.

In practice, this affects http: right now, but this patch also ensures
that the "omit https" and "omit file" bits are also not set to be
more futureproof.

R=npm@chromium.org

Bug: 1468681
Change-Id: Idac394102133f030dad899e9017fb5e3f0cb4b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4742567
Reviewed-by: Nicolás Peña <npm@chromium.org>
Commit-Queue: Christian Biesinger <cbiesinger@chromium.org>
Auto-Submit: Christian Biesinger <cbiesinger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1179216}
NOKEYCHECK=True
GitOrigin-RevId: 4352cfb7ea41c9051f14c1a8d90d7ab487a9bc59