Reland "Add WebDriver test for CSS 'resize'"
This is a reland of commit 488fd2c958c9c4a5e38307798f2a1ac25559977e

The commit was reverted because the test was failing when run with
--enable-leak-detection, with this leak log:
  ({
    "numberOfLiveDocuments":[1,2],
    "numberOfLiveNodes":[4,37],
    "numberOfLiveResourceFetchers":[1,2],
    "numberOfLiveResources":[0,5]
  })

I think that's a problem of the WebDriver API.

Removing the test cases trying to resize an element with 'resize: none'
avoids the problem, and since that's the initial value, they don't seem
much important.

So I think we can reland the test without these cases.

Original change's description:
> Add WebDriver test for CSS 'resize'
>
> In particular, this tests the change from CL:4308393, i.e. resolving
> 'resize: block' and 'resize: inline' using the writing mode of the
> element itself, instead of the one of the containing block.
>
> Bug: 850004
> Change-Id: I15deae5cabd5e92afd6d0405f46769be525c0bd2
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4738818
> Reviewed-by: Rune Lillesveen <futhark@chromium.org>
> Commit-Queue: Oriol Brufau <obrufau@igalia.com>
> Cr-Commit-Position: refs/heads/main@{#1178430}

Bug: 850004
Change-Id: Ifd4196b15263adb5dd71cc5c83e264a4812e2fa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4742606
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Oriol Brufau <obrufau@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1179390}
NOKEYCHECK=True
GitOrigin-RevId: e84540c198a49e58c80f094d90ecf99a8d226bf4