[text-autospace] Implement spacing insertion logic in ShapeResult
The callers are supposed to compute the positions where the spacing
should be inserted to, with the width of the spacing, and then pass the
list to ShapeResult.
The ShapeResult will scan glyph_data_ and add extra spacing.

A follow-up will test and implement the RTL case.

Bug: 1463890
Change-Id: I1ca40c15047faa5dc840d29fced0043a3eefbf40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4745080
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1179448}
NOKEYCHECK=True
GitOrigin-RevId: 20519fbcfe2704a7bbb59f9465853e15a5fb7a3c