Use moved GL GrBackendSurface functions/methods
In https://skia-review.googlesource.com/c/skia/+/701398 we are refactoring GrBackend* to not require #ifdefs. This changes
callsites in Chromium to use static functions instead of
methods that were conditionally compiled on those classes.

There should be no functional change with these. As I was updating
these callsites, I saw uses of the deprecated GrMipMapped enum
and used the updated skgpu::Mipmapped (the former is an alias
to the latter).

Bug: b/293490566
Change-Id: I0373db0eae464ea4dff6d218be8df311d1e2bdd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4739911
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Kevin Lubick <kjlubick@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1179907}
NOKEYCHECK=True
GitOrigin-RevId: b9a3de1ceb823eca4eb3d46d8ccc308edc932d52