Add enterprise policy to control whether unload can be deprecated.
This adds a command-line switch
unload-event-handler-deprecation-disallowed that is set on every
renderer if deprecation is allowed.

One tricky point is that it is never set correctly on the browser
process. I think this is OK because the only use for the feature list in the browser is for fenced frames and fenced frames never allow
unload. However this seems like a possible trap in the future. One way
to clean that up would be to stop special casing fenced-frames and
unload in EventTarget::AddEventListenerInternal and instead add a
GetPermissionsPolicyFeatureListForFencedFrame now that unload is
controlled by PP.

Bug: 1423766,1432116
Change-Id: I5489519a74ef16da2af4e4c2669a31c3a0577b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4730081
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Fergal Daly <fergal@chromium.org>
Reviewed-by: Victor Gabriel Savu <vsavu@google.com>
Cr-Commit-Position: refs/heads/main@{#1180209}
NOKEYCHECK=True
GitOrigin-RevId: 0045f193e2fdeb4cba8b53269cc1b1ef0f29b68f