URL: Measure setting a non-special scheme on a special scheme URL
Chrome currently allows setting a non-special scheme on a special
schema URL. This is not compliant with the URL standard.

Example:

    const url = new URL("https://example.com/");
    url.protocol = "non-special";
    // url.href should be "https://example.com/",
    // but it became "non-special://example.com/"

To address this interoperability issue, we need to assess the impact
of fixing it.

This is in preparation of the main CL [1] and Intent to Ship, which
will be sent.

[1]: https://crrev.com/c/4744799

Bug: 1416018
Change-Id: I0304bb673302b127f35aef598b566465800d888b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4745821
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Hayato Ito <hayato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1180604}
NOKEYCHECK=True
GitOrigin-RevId: 54fc40d6622f38b3071b1f73bf403ed9d38d2a99